+++ CHANGELOG.md
@@ -13,6 +13,8 @@ Wherever possible, a reference to an issue in the project issue tracker should b
 
 * Refactored Data Catalogue exporter with support for catalogue items and other derived formats
   [#137](https://gitlab.data.bas.ac.uk/MAGIC/assets-tracking-service/-/issues/137)
+* RVDAS provider for tracking the SDA
+  [#25](https://gitlab.data.bas.ac.uk/MAGIC/assets-tracking-service/-/issues/25)
 
 ## [0.5.0] - 2025-03-15